refactor: fake 객체 패키지 분리
